+++ src/features/experience/Confetti.tsx
@@ -1,4 +1,6 @@
-import { useCallback, useRef } from 'react';
+'use client';
+
+import { useCallback, useEffect, useRef } from 'react';
 import ReactCanvasConfetti from 'react-canvas-confetti';
 
 const Confetti = () => {
@@ -7,6 +9,48 @@ const Confetti = () => {
   const getInstance = useCallback((instance: any) => {
     refAnimationInstance.current = instance;
   }, []);
+
+  const makeShot = useCallback((particleRatio: number, opts: object) => {
+    refAnimationInstance.current &&
+      refAnimationInstance.current({
+        ...opts,
+        origin: { y: 0.7 },
+        particleCount: Math.floor(200 * particleRatio),
+      });
+  }, []);
+
+  const fire = useCallback(() => {
+    makeShot(0.25, {
+      spread: 26,
+      startVelocity: 55,
+    });
+
+    makeShot(0.2, {
+      spread: 60,
+    });
+
+    makeShot(0.35, {
+      spread: 100,
+      decay: 0.91,
+      scalar: 0.8,
+    });
+
+    makeShot(0.1, {
+      spread: 120,
+      startVelocity: 25,
+      decay: 0.92,
+      scalar: 1.2,
+    });
+
+    makeShot(0.1, {
+      spread: 120,
+      startVelocity: 45,
+    });
+  }, [makeShot]);
+
+  useEffect(() => {
+    fire();
+  }, [fire]);
 
   return (
     <ReactCanvasConfetti
